# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Credentials": {
        "additionalProperties": false,
        "description": "认证凭据。**绝不**写进日志：观测层只记 sha256 前 12 位。",
        "properties": {
          "scheme": {
            "default": "bearer",
            "title": "Scheme",
            "type": "string"
          },
          "token": {
            "maxLength": 512,
            "minLength": 8,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "Credentials",
        "type": "object"
      },
...
      "PrincipalDTO": {
        "description": "主体声明（OpenAPI 组件）。",
        "properties": {
          "roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Roles"
          },
          "subject": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Subject"
          }
        },
        "title": "PrincipalEnvelope",
        "type": "object"
      },
      "RetrieveRequest": {
        "description": "POST /v1/knowledge/retrieve 的请求体。",
        "properties": {
          "api_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Api Version"
          },
          "budget_ms": {
            "anyOf": [
              {
                "maximum": 600000,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Budget Ms"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextDTO"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "credentials": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Credentials"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "decision_ref": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Decision Ref"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Idempotency Key"
          },
          "limit": {
            "anyOf": [
              {
                "maximum": 50,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Limit"
          },
          "principal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrincipalDTO"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "query": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Query"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "tenant": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tenant"
          },
          "trace_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Trace Id"
          },
          "workspace": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Workspace"
          }
        },
        "title": "RetrieveRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Authorization: Bearer <token>；令牌只以 sha256 存在配置里",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "把规则、检索与验证能力服务化；判定仍由核心 Policy Engine 做。",
    "title": "Engineering Policy Platform API",
    "version": "1.0",
    "x-api-schema-version": "1.0",
    "x-decision-schema-version": "1.1",
    "x-policy-generation": "decision-1.1",
    "x-service-version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/knowledge/retrieve": {
      "post": {
        "description": "检索离线规范并组装带来源的 Engineering Context。",
        "operationId": "knowledgeRetrieve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "决策 / 检索结果 / 验证证据（载荷是受控 JSON）"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "400：结构化错误（见 ErrorResponse）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "401：结构化错误（见 ErrorResponse）"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "403：结构化错误（见 ErrorResponse）"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "404：结构化错误（见 ErrorResponse）"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "409：结构化错误（见 ErrorResponse）"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "413：结构化错误（见 ErrorResponse）"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "415：结构化错误（见 ErrorResponse）"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "429：结构化错误（见 ErrorResponse）"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "500：结构化错误（见 ErrorResponse）"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "503：结构化错误（见 ErrorResponse）"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "504：结构化错误（见 ErrorResponse）"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retrieve",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/v1/policy/evaluate": {
      "post": {
        "description": "对固定上下文给出 allow / allow_with_warnings / block。",
        "operationId": "policyEvaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "决策 / 检索结果 / 验证证据（载荷是受控 JSON）"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "400：结构化错误（见 ErrorResponse）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "401：结构化错误（见 ErrorResponse）"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "403：结构化错误（见 ErrorResponse）"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "404：结构化错误（见 ErrorResponse）"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "409：结构化错误（见 ErrorResponse）"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "413：结构化错误（见 ErrorResponse）"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "415：结构化错误（见 ErrorResponse）"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "429：结构化错误（见 ErrorResponse）"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "500：结构化错误（见 ErrorResponse）"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "503：结构化错误（见 ErrorResponse）"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "504：结构化错误（见 ErrorResponse）"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Evaluate",
        "tags": [
          "policy"
        ]
      }
    },
    "/v1/validation/evaluate": {
      "post": {
        "description": "产出验证器证据（可选顺带判定）。",
        "operationId": "validationEvaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "决策 / 检索结果 / 验证证据（载荷是受控 JSON）"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "400：结构化错误（见 ErrorResponse）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "401：结构化错误（见 ErrorResponse）"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "403：结构化错误（见 ErrorResponse）"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "404：结构化错误（见 ErrorResponse）"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "409：结构化错误（见 ErrorResponse）"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "413：结构化错误（见 ErrorResponse）"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "415：结构化错误（见 ErrorResponse）"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "429：结构化错误（见 ErrorResponse）"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "500：结构化错误（见 ErrorResponse）"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "503：结构化错误（见 ErrorResponse）"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "504：结构化错误（见 ErrorResponse）"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Validate",
        "tags": [
          "validation"
        ]
      }
    }
  }
}